Delete: login app, Modify: users.view, urls / westagram.urls
