--- students/byoungju/login/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from login.views import LoginView
-
-urlpatterns = [
-    path("", LoginView.as_view()),
-]
