+++ students/byoungju/users/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
-from users.views import UserView
+from users.views import UserView, LoginView
 
 urlpatterns = [
     path("", UserView.as_view()),
+    path("/login", LoginView.as_view()),
 ]
