+++ students/byoungju/users/views.py
@@ -34,5 +34,22 @@ class UserView(View):
         except KeyError:
             return JsonResponse({"message":"KEY_ERROR"}, status = 400)
 
-            
+class LoginView(View):
+    def post(sef, request):
+        data = json.loads(request.body)
+ 
+        try:
+            if (data["email"] == "" or data["password"] == ""):
+                return JsonResponse ({"message":"INVALID_USER"}, status = 401)
+
+            if (User.objects.filter(email = data["email"]).exists() == False):
+                return JsonResponse({"message":"INVALID_USER"}, status = 401)
+
+            if (User.objects.get(email = data["email"]).password != data["password"]):
+                return JsonResponse({"message":"INVALID_USER"}, status = 401)
+
+            return JsonResponse({"message":"SUCCESS"}, status = 200)
+
+        except KeyError:
+            return JsonResponse({"message":"KEY_ERROR"}, status = 400)
