+++ students/byoungju/westagram/urls.py
@@ -2,5 +2,4 @@ from django.urls import path, include
 
 urlpatterns = [
     path("users", include("users.urls")),
-    path("login", include("login.urls")),
 ]
